+++ src/components/Navbar/web/Navbar.js
@@ -6,17 +6,17 @@ import { NavbarButton } from "./NavbarButton";
 export function Navbar() {
   return (
     // sticky top-0 z-10 to keep at top
-    <div className="h-16 flex justify-center duration-500 bg-gradient-to-b from-blue-300 to-blue-200">
+    <div className="flex duration-500 bg-gradient-to-b from-blue-300 to-blue-200 justify-between pr-5">
+      <Link to="/">
+        <div className="w-10 h-10 m-3 hover:w-16 hover:h-16 duration-1000">
+          <img src={logo} alt="Mark" className="rounded-full " />
+        </div>
+      </Link>
       <div className="flex">
         <NavbarButton title="HOME" link="/" />
-        <NavbarButton title="RESUME" link="/resume" />
-        <Link to="/">
-          <div className="w-14 h-14 m-1">
-            <img src={logo} alt="Mark" className="rounded-full" />
-          </div>
-        </Link>
         <NavbarButton title="PROJECTS" link="/projects" />
         <NavbarButton title="MUSIC" link="/music" />
+        <NavbarButton title="RESUME" link="/resume" />
       </div>
     </div>
   );