+++ src/components/homepage/SectionAboutMe.js
@@ -2,9 +2,9 @@ import email from "../../assets/icons/email.png";
 import instagram from "../../assets/icons/instagram.png";
 import github from "../../assets/icons/github.png";
 import linkedin from "../../assets/icons/linkedin.png";
-import { ShellComponent } from "./ShellComponent";
 import data from "../../info.json";
 import { useState, useEffect } from "react";
+import ProfilePic from "../../assets/profile_pic.jpg";
 
 export function SectionAboutMe() {
   const contact = data.contact;
@@ -32,49 +32,53 @@ export function SectionAboutMe() {
   }, []);
 
   return (
-    <div className="my-5">
-      <div className="md:grid md:grid-cols-2">
-        <div className="flex justify-center items-center">
-          <div className="mt-10 mb-16 md:my-0">
-            <h1 className="cursor-pointer text-black font-light text-6xl text-center">
-              {typedText || <span className="invisible">Mark Li</span>}
-            </h1>
-            <div className="flex gap-6 justify-center items-center mt-3">
-              <div
-                onClick={() => {
-                  navigator.clipboard.writeText(contact.email);
-                }}
-                style={{ cursor: "pointer", position: "relative" }}
-                title="Click to copy email"
-              >
-                <img src={email} alt="email" className="w-8 h-8" />
-              </div>
-              <a
-                href={contact.instagram}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <img src={instagram} alt="instagram" className="w-8 h-8" />
-              </a>
-              <a
-                href={contact.github}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <img src={github} alt="github" className="w-8 h-8" />
-              </a>
-              <a
-                href={contact.linkedin}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                <img src={linkedin} alt="linkedin" className="w-8 h-8" />
-              </a>
+    <div className="my-20 mx-4">
+      <div className="flex justify-center items-center duration-300">
+        <img
+          src={ProfilePic}
+          alt="profile"
+          className="sm:w-64 sm:h-64 w-40 h-40 duration-300 sm:mr-8 mr-4 rounded-full object-cover"
+        />
+        <div>
+          <h1 className="text-black font-light text-5xl sm:text-6xl duration-300">
+            {typedText || <span className="invisible">Mark Li</span>}
+          </h1>
+          <h2 className="ml-1 font-light text-md sm:text-lg mt-1">
+            <span className="text-gray-500">Software Developer</span>
+            <br />
+            <span className="text-gray-500">Amateur Musician</span>
+          </h2>
+          <div className="flex gap-4 sm:gap-6 space-between mt-2 ml-1 duration-300">
+            <div
+              onClick={() => {
+                navigator.clipboard.writeText(contact.email);
+              }}
+              style={{ cursor: "pointer", position: "relative" }}
+              title="Click to copy email"
+            >
+              <img src={email} alt="email" className="w-8 h-8" />
             </div>
+            <a
+              href={contact.instagram}
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <img src={instagram} alt="instagram" className="w-8 h-8" />
+            </a>
+            <a href={contact.github} target="_blank" rel="noopener noreferrer">
+              <img src={github} alt="github" className="w-8 h-8" />
+            </a>
+            <a
+              href={contact.linkedin}
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <img src={linkedin} alt="linkedin" className="w-8 h-8" />
+            </a>
           </div>
         </div>
-        <ShellComponent />
       </div>
+      {/* <AboutMe /> */}
     </div>
   );
 }