+++ src/pages/Home.js
@@ -6,7 +6,7 @@ import {
 
 export function Home() {
   return (
-    <div className="bg-gradient-to-b from-blue-200 via-white to-red-50 py-10">
+    <div className="bg-gradient-to-b from-blue-200 via-white to-red-100 py-10">
       <SectionAboutMe />
       <SectionProjects />
       <SectionMusic />